Fixing VirtualNumbers API and test. Fixing ContactsUpload API and test.
