+++ src/YouMailAPI/Virtual Numbers/VirtualNumbersAPI.cs
@@ -1,5 +1,5 @@
 /*************************************************************************************************
- * Copyright (c) 2018 Gilles Khouzam
+ * Copyright (c) 2021 Gilles Khouzam
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy of this software
  * and associated documentation files (the "Software"), to deal in the Software withou
@@ -17,23 +17,36 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 /*************************************************************************************************/
 
-namespace MagikInfo.YouMailAPI.Tests
+namespace MagikInfo.YouMailAPI
 {
-    using Microsoft.VisualStudio.TestTools.UnitTesting;
-    using System;
+    using System.Net.Http;
     using System.Threading.Tasks;
 
-    [TestClass]
-    public class YouMailVirtualNumberTests
+    public partial class YouMailService
     {
-        YouMailService Service = YouMailTestService.MyService;
-
-        [TestMethod]
-        public async Task GetVirtualNumbers()
+        /// <summary>
+        /// Get the list of virtual numbers for the user
+        /// </summary>
+        /// <returns></returns>
+        public async Task<YouMailVirtualNumber[]> GetVirtualNumbersAsync()
         {
-            var numbers = await Service.GetVirtualNumbersAsync();
-            Assert.IsNotNull(numbers);
-            Assert.IsTrue(numbers.Length >= 1);
+            YouMailVirtualNumber[] returnValue = null;
+            try
+            {
+                AddPendingOp();
+                if (await LoginWaitAsync())
+                {
+                    using (var response = await YouMailApiAsync(YMST.c_virtualNumbersSettings, null, HttpMethod.Get))
+                    {
+                        returnValue = DeserializeObject<YouMailVirtualNumber[]>(response.GetResponseStream(), YMST.c_virtualNumberSettingsList);
+                    }
+                }
+                return returnValue;
+            }
+            finally
+            {
+                RemovePendingOp();
+            }
         }
     }
 }
